fix: seed E2E fixtures in-process instead of through a deployed edge function
The Web E2E job seeded its fixture identity by POSTing to a deployed
`e2e-fixtures` Supabase Edge Function. That put the caller and the seeder on
independent release cycles: the caller shipped with its pull request, while the
seeder only changed when somebody remembered to run `supabase functions
deploy`. This branch's caller started sending the exact run-scoped identity it
wanted seeded, the deployed copy still derived its own, and the specs signed in
as a user that was never created. `auth-shell.spec.ts` and
`profile-completion.spec.ts` then timed out waiting for a URL they could never
reach.

Seeding now runs in the Node process the specs already spawn from `beforeEach`,
against the Supabase admin API with the service-role key. The client versus
deployed-server skew class is gone: there is one seeding implementation, in the
repository, versioned with its caller.

Changes:

- `apps/web-console/tests/e2e/support/e2e-fixture-seed.mjs` (new) holds all
  service-role work, ported from the edge function: the three fixture users,
  tenants, accounts, memberships, profiles, the pending invitation, and the
  stale-run sweep. It also keeps the run-key derivation this branch added, now
  in the only place that derives it.
- `e2e-auth-fixtures.mjs` resolves the credentials and passes them straight to
  the seeder, so one process decides the identity and writes it. It also gained
  a `reset-profile <email>` action so the CLI is the single entrypoint.
- `reset-profile.ts` shells out to that action instead of POSTing to the edge
  function. Shelling out rather than importing avoids Playwright's CommonJS
  transform choking on a .mjs import, and keeps the key out of the Playwright
  worker's module graph.
- Seeding without `SUPABASE_URL` and `SUPABASE_SERVICE_ROLE_KEY` now fails and
  names the missing variable. It used to silently no-op, which is what let the
  skew hide until a 25 second sign-in timeout.
- `supabase/functions/e2e-fixtures/` is deleted. Nothing else called it, and
  leaving a second seeding implementation in the tree is what drifted.
- `ci.yml` drops `E2E_FIXTURE_URL` and `E2E_FIXTURE_SECRET` from the Playwright
  step and scopes `SUPABASE_SERVICE_ROLE_KEY` to it instead. The key stays out
  of job-level env, matching the other two steps that need admin auth. The
  repository secrets themselves are left in place.
- Passwords now come from the same `E2E_VERIFIED_PASSWORD` and
  `E2E_UNVERIFIED_PASSWORD` the specs read. The edge function read a separate
  `E2E_DEFAULT_*` pair that CI never set, which was a second way for the two
  sides to disagree.
- `tests/unit/e2e-fixture-ids.test.ts` (new) covers the run-key derivation and
  the secret redaction that were previously unreachable from any test in this
  repository.
