+++ apps/web-console/tests/e2e/support/reset-profile.ts
@@ -1,42 +1,31 @@
+import { execFileSync } from "node:child_process";
 import { E2E_VERIFIED_EMAIL } from "./e2e-auth-creds";
 
-// Spec-side twin of `resetProfileBetweenSpecs` in `e2e-auth-fixtures.mjs`.
-// Lives in a .ts module because Playwright's transform compiles spec imports
-// to CommonJS, and importing a .mjs from a spec makes Node evaluate that
-// CJS output in ES module scope ("exports is not defined"). The .mjs copy
-// remains for CLI (child process) callers only.
-export async function resetProfileBetweenSpecs(testInfo?: {
-  email?: string;
-}): Promise<unknown | null> {
-  const url = process.env.E2E_FIXTURE_URL;
-  const secret = process.env.E2E_FIXTURE_SECRET;
-  if (!url || !secret) {
-    return null;
-  }
+// Spec-side entry to the fixture CLI's `reset-profile` action.
+//
+// Shells out rather than importing the seeder: Playwright's transform compiles
+// spec imports to CommonJS, and importing a .mjs from a spec makes Node
+// evaluate that CJS output in ES module scope ("exports is not defined"). The
+// child process also keeps this identical to the full reset the beforeEach
+// hooks already run, and keeps the service-role key out of the Playwright
+// worker's own module graph.
+export function resetProfileBetweenSpecs(testInfo?: { email?: string }): void {
   const targetEmail = testInfo?.email ?? E2E_VERIFIED_EMAIL;
-  const response = await fetch(url, {
-    method: "POST",
-    headers: {
-      "X-E2E-Secret": secret,
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      action: "reset-profile",
-      email: targetEmail,
-    }),
-  });
-  const text = await response.text();
-  let data: unknown = null;
   try {
-    data = text ? JSON.parse(text) : null;
-  } catch {
-    data = { raw: text };
+    execFileSync(
+      "node",
+      ["tests/e2e/support/e2e-auth-fixtures.mjs", "reset-profile", targetEmail],
+      {
+        cwd: process.cwd(),
+        env: { ...process.env, NODE_OPTIONS: "" },
+        stdio: "pipe",
+      }
+    );
+  } catch (err: unknown) {
+    const e = err as { stdout?: Buffer; stderr?: Buffer };
+    process.stderr.write(
+      `[e2e-auth-fixtures] reset-profile failed\n${e.stdout ?? ""}${e.stderr ?? ""}\n`
+    );
+    throw err;
   }
-  if (!response.ok) {
-    const message =
-      (data as { error?: string } | null)?.error ??
-      `${response.status} ${response.statusText}`;
-    throw new Error(`resetProfileBetweenSpecs failed: ${message}`);
-  }
-  return data;
 }
